Move to java 17 but check that LJC can convert the lib back to java 11. Also add fingerprint support.

+++ Server/src/main/java/org/moddingx/cursewrapper/backend/data/response/ModFileResponse.java
@@ -31,7 +31,7 @@ public class ModFileResponse implements CurseData {
         @Expose public FileStatus fileStatus;
         @Expose public List<FileHash> hashes;
         @Expose public Date fileDate;
-        @Expose public int fileLength;
+        @Expose public long fileLength;
         @Expose public int downloadCount;
         @Expose public String downloadUrl;
         @Expose public List<String> gameVersions;
